+++ client/src/main/scala/sbt/client/impl/DebugClient.scala
@@ -8,7 +8,7 @@ import sbt.protocol.RegisterClientRequest
 import sbt.protocol.ClientInfo
 import sbt.protocol.registerClientRequestWrites
 
-object DebugClient {
+private[sbt] object DebugClient {
   def apply(port: Int): SbtClient = {
     val client = new ipc.Client(new Socket("127.0.0.1", port))
     val uuid = java.util.UUID.randomUUID()
