+++ commons/protocol/src/main/scala/sbt/protocol/ScalaShims.scala
@@ -1,5 +1,5 @@
 package sbt.protocol
 
-object ScalaShims {
+private[sbt] object ScalaShims {
   val ManifestFactory = scala.reflect.ManifestFactory
 }
